Format

+++ src/KmanDEXPool.sol
@@ -1,16 +1,18 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.28;
+
 import "../lib/forge-std/src/console.sol";
 
 contract KmanDEXPool {
-
     constructor(address tokenA, address tokenB) {
         // Initialize the pair with tokenA and tokenB
         console.log("Pair created with tokens:", tokenA, tokenB);
-
     }
 
-    function addLiquidity(address tokenA, address tokenB, uint amountTokenA, uint amountTokenB) external payable {
+    function addLiquidity(address tokenA, address tokenB, uint256 amountTokenA, uint256 amountTokenB)
+        external
+        payable
+    {
         // Add liquidity logic
         console.log("Adding liquidity");
     }
